+++ backup/build.bak
@@ -44,23 +44,23 @@ def gen_blog():
         template = template.replace("{{blog}}", partial)
         open(p["output"], "w+").write(template)
 
-#Jinja stuff, Work in progress
-def read_template():
+#Jinja stuff, Work in progress, getting an error....
+def fix_template():
 
     for p in pages:
-        index_html = open(p["pages"]).read() 
-        template_html = open("templates/base.html").read() 
-        template = Template(template_html)
-        template = template.replace("{{title}}", p["title"])
+        index_html = open(p["filename"]).read() 
+        #template_html = open("templates/base.html").read() 
+        template = Template(index_html)
+        template = template.render({'title':p["title"]})
         open(p["output"], "w+").write(template)
 
 
 
 def main():
     read_files()
-    #read_template()
     gen_html()
-    #gen_blog()
+    gen_blog()
+    #fix_template()
     print("Your files have been generated")
 
 if __name__ == "__main__":
