+++ build.py
@@ -30,8 +30,6 @@ def gen_html():
     for p in pages:
         template = open("./templates/base.html").read()
         partial = open(p["filename"]).read()
-        #template = template.replace("{{title}}", p["title"])
-        #template = template.replace("{{content}}", partial)
         open(p["output"], "w+").write(template)
 
 
@@ -39,28 +37,25 @@ def gen_blog():
 
     for b in blog:
         template = open("./templates/blog.html").read()
-        partial = open(p["filename"]).read()
-        #template = template.replace("{{title}}", p["title"])
-        #template = template.replace("{{blog}}", partial)
-        open(p["output"], "w+").write(template)
+        partial = open(b["filename"]).read()
+        open(b["output"], "w+").write(template)
 
-def fix_templates():
+def fix_template():
 
     for p in pages:
         partial = open(p["filename"]).read() 
         template_html = open("templates/base.html").read()
         template = Template(template_html)
+        #template = template.render({'title': p['title']})
         template = template.render({'title': p['title'], 'content': partial})
-        #template = template.render({'content': partial})
         open(p["output"], "w+").write(template)
 
     for b in blog:
-        partial = open(b["filename"]).read() 
-        blog_template = open("templates/blog.html").read()
-        template = Template(blog_template)
+        partial = open(p["filename"]).read() 
+        template_html = open("templates/base.html").read()
+        template = Template(template_html)
         #template = template.render({'title': b['title']})
-        template = template.render({'title': b['title'], 'blog': partial})
-        #template = template.render({'blog': b['partial']})
+        template = template.render({'title': p['title'], 'blog': partial})
         open(b["output"], "w+").write(template)
 
 
@@ -69,7 +64,7 @@ def main():
     read_files()
     gen_html()
     gen_blog()
-    fix_templates()
+    fix_template()
     print("Your files have been generated")
 
 if __name__ == "__main__":
