+++ build_borked.py
@@ -30,36 +30,46 @@ def gen_html():
     for p in pages:
         template = open("./templates/base.html").read()
         partial = open(p["filename"]).read()
-        template = template.replace("{{title}}", p["title"])
-        template = template.replace("{{content}}", partial)
+        #template = template.replace("{{title}}", p["title"])
+        #template = template.replace("{{content}}", partial)
         open(p["output"], "w+").write(template)
 
 
 def gen_blog():
 
-    for p in blog:
+    for b in blog:
         template = open("./templates/blog.html").read()
         partial = open(p["filename"]).read()
-        template = template.replace("{{title}}", p["title"])
-        template = template.replace("{{blog}}", partial)
+        #template = template.replace("{{title}}", p["title"])
+        #template = template.replace("{{blog}}", partial)
         open(p["output"], "w+").write(template)
 
-#Jinja stuff, Work in progress
-def read_template():
+def fix_templates():
 
-    index_html = open("index.html").read()
-    template_html = open("base.html").read()
-    template = Template(template_html)
-    template.render(title="Homepage",content=index.html)
+    for p in pages:
+        partial = open(p["filename"]).read() 
+        template_html = open("templates/base.html").read()
+        template = Template(template_html)
+        template = template.render({'title': p['title'], 'content': partial})
+        #template = template.render({'content': partial})
+        open(p["output"], "w+").write(template)
 
+    for b in blog:
+        partial = open(b["filename"]).read() 
+        blog_template = open("templates/blog.html").read()
+        template = Template(blog_template)
+        #template = template.render({'title': b['title']})
+        template = template.render({'title': b['title'], 'blog': partial})
+        #template = template.render({'blog': b['partial']})
+        open(b["output"], "w+").write(template)
 
 
 
 def main():
     read_files()
-    #read_template()
     gen_html()
-    #gen_blog()
+    gen_blog()
+    fix_templates()
     print("Your files have been generated")
 
 if __name__ == "__main__":
