+++ docs/blog.html
@@ -33,7 +33,12 @@
 				<div class="col-sm-8">
 					<div class="posts">
 						<div>
-							
+													<h1>Kickstart coding, Week 1</h1>
+						<p>10/25/19</p>
+						<p>This week I started my journey at Kickstart coding Bootcamp in Oakland. It is a full stack bootcamp.</p>
+						<p>Some of the technologies we will be learking are: HTML, CSS, JS, Python, Django and much more.</p>
+						<p>The first week we covered html, css and bootstrap.</p><p>Our home work assignment this week is to create our first website/portfolio.</p>
+
 						</div>
 					</div>
 				</div> <!-- end col-sm-8 -->
