+++ docs/blog2.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 	<head>
-    <title>{{title}}</title>
+    <title>Projects</title>
   		<meta charset="utf-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1">
 		<link href="https://fonts.googleapis.com/css?family=PT+Sans&display=swap" rel="stylesheet"> 
@@ -9,7 +9,6 @@
 		<link rel="stylesheet" href="./css/bootstrap.css">
 	</head>
 	<body>
-
 		<div class="container" style="margin-top:30px">
 			<div class="row">
 				<div class="col-sm-4">
@@ -21,16 +20,7 @@
 							<a class="nav-link active" href="index.html">Home</a>
 						</li>
 						<li class="nav-item">
-							<a class="nav-link" href="blog1.html">Week1</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="blog2.html">Week2</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="blog3.html">Week3</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="blog4.html">Week4</a>
+							<a class="nav-link" href="blog1.html">Blog</a>
 						</li>
 						<li class="nav-item">
 							<a class="nav-link" href="projects.html">Projects</a>
@@ -40,12 +30,12 @@
 						</li>
 					</ul>
 				</div> <!-- end col-sm-4 -->
-        		<div class="col-sm-8">
+				<div class="col-sm-8">
 					<div class="posts">
 						<div>
-           					{{blog}}
+							
 						</div>
-					</div> <!-- end posts -->
+					</div>
 				</div> <!-- end col-sm-8 -->
 			</div> <!-- end row -->
 		</div>
